requirements.txt
Flask==2.2.2
mysql-connector-python==8.0.29
gunicorn==20.1.0